requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
deptry==0.12.0
Django==5.0.1
djangorestframework==3.14.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
